requirements.txt
tensorflow-cpu==2.12.0
numpy==1.23.5
discord==2.2.2
scikit-learn==1.2.2
